backend/requirements.txt
fastapi[standard]>=0.115.0
pydantic>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
